io_service.hpp: use runtime feature detection

+++ io_service.hpp
@@ -15,10 +15,6 @@
 #include "promise.hpp"
 #include "task.hpp"
 
-#ifndef LINUX_KERNEL_VERSION
-#   define LINUX_KERNEL_VERSION 58
-#endif
-
 /** Fill an iovec struct using buf & size */
 constexpr inline iovec to_iov(void *buf, size_t size) noexcept {
     return { buf, size };
@@ -117,64 +113,70 @@ public:
 
         io_uring_queue_init_params(entries, &ring, &p) | panic_on_err("queue_init_params", false);
 
-#ifndef NDEBUG
         auto* probe = io_uring_get_probe_ring(&ring);
         on_scope_exit free_probe([=]() { io_uring_free_probe(probe); });
-#   define TEST_IORING_OP(opcode) do { for (int i = 0; i < probe->ops_len; ++i) { if (probe->ops[i].op == opcode && probe->ops[i].flags & IO_URING_OP_SUPPORTED) { puts("\t" #opcode); break; } } } while (0)
-        puts("Supported io_uring opcodes by current kernel:");
-        TEST_IORING_OP(IORING_OP_NOP);
-        TEST_IORING_OP(IORING_OP_READV);
-        TEST_IORING_OP(IORING_OP_WRITEV);
-        TEST_IORING_OP(IORING_OP_FSYNC);
-        TEST_IORING_OP(IORING_OP_READ_FIXED);
-        TEST_IORING_OP(IORING_OP_WRITE_FIXED);
-        TEST_IORING_OP(IORING_OP_POLL_ADD);
-        TEST_IORING_OP(IORING_OP_POLL_REMOVE);
-        TEST_IORING_OP(IORING_OP_SYNC_FILE_RANGE);
-        TEST_IORING_OP(IORING_OP_SENDMSG);
-        TEST_IORING_OP(IORING_OP_RECVMSG);
-        TEST_IORING_OP(IORING_OP_TIMEOUT);
-        TEST_IORING_OP(IORING_OP_TIMEOUT_REMOVE);
-        TEST_IORING_OP(IORING_OP_ACCEPT);
-        TEST_IORING_OP(IORING_OP_ASYNC_CANCEL);
-        TEST_IORING_OP(IORING_OP_LINK_TIMEOUT);
-        TEST_IORING_OP(IORING_OP_CONNECT);
-        TEST_IORING_OP(IORING_OP_FALLOCATE);
-        TEST_IORING_OP(IORING_OP_OPENAT);
-        TEST_IORING_OP(IORING_OP_CLOSE);
-        TEST_IORING_OP(IORING_OP_FILES_UPDATE);
-        TEST_IORING_OP(IORING_OP_STATX);
-        TEST_IORING_OP(IORING_OP_READ);
-        TEST_IORING_OP(IORING_OP_WRITE);
-        TEST_IORING_OP(IORING_OP_FADVISE);
-        TEST_IORING_OP(IORING_OP_MADVISE);
-        TEST_IORING_OP(IORING_OP_SEND);
-        TEST_IORING_OP(IORING_OP_RECV);
-        TEST_IORING_OP(IORING_OP_OPENAT2);
-        TEST_IORING_OP(IORING_OP_EPOLL_CTL);
-        TEST_IORING_OP(IORING_OP_SPLICE);
-        TEST_IORING_OP(IORING_OP_PROVIDE_BUFFERS);
-        TEST_IORING_OP(IORING_OP_REMOVE_BUFFERS);
-        TEST_IORING_OP(IORING_OP_TEE);
-        TEST_IORING_OP(IORING_OP_SHUTDOWN);
-        TEST_IORING_OP(IORING_OP_RENAMEAT);
-        TEST_IORING_OP(IORING_OP_UNLINKAT);
-        TEST_IORING_OP(IORING_OP_MKDIRAT);
-#   undef TEST_IORING_OP
+#define TEST_IORING_OP(opcode) do {\
+    for (int i = 0; i < probe->ops_len; ++i) {\
+        if (probe->ops[i].op == opcode && probe->ops[i].flags & IO_URING_OP_SUPPORTED) {\
+            probe_ops[i] = true;\
+            puts("\t" #opcode);\
+            break;\
+            }\
+        }\
+    } while (0)
+    puts("Supported io_uring opcodes by current kernel:");
+    TEST_IORING_OP(IORING_OP_NOP);
+    TEST_IORING_OP(IORING_OP_READV);
+    TEST_IORING_OP(IORING_OP_WRITEV);
+    TEST_IORING_OP(IORING_OP_FSYNC);
+    TEST_IORING_OP(IORING_OP_READ_FIXED);
+    TEST_IORING_OP(IORING_OP_WRITE_FIXED);
+    TEST_IORING_OP(IORING_OP_POLL_ADD);
+    TEST_IORING_OP(IORING_OP_POLL_REMOVE);
+    TEST_IORING_OP(IORING_OP_SYNC_FILE_RANGE);
+    TEST_IORING_OP(IORING_OP_SENDMSG);
+    TEST_IORING_OP(IORING_OP_RECVMSG);
+    TEST_IORING_OP(IORING_OP_TIMEOUT);
+    TEST_IORING_OP(IORING_OP_TIMEOUT_REMOVE);
+    TEST_IORING_OP(IORING_OP_ACCEPT);
+    TEST_IORING_OP(IORING_OP_ASYNC_CANCEL);
+    TEST_IORING_OP(IORING_OP_LINK_TIMEOUT);
+    TEST_IORING_OP(IORING_OP_CONNECT);
+    TEST_IORING_OP(IORING_OP_FALLOCATE);
+    TEST_IORING_OP(IORING_OP_OPENAT);
+    TEST_IORING_OP(IORING_OP_CLOSE);
+    TEST_IORING_OP(IORING_OP_FILES_UPDATE);
+    TEST_IORING_OP(IORING_OP_STATX);
+    TEST_IORING_OP(IORING_OP_READ);
+    TEST_IORING_OP(IORING_OP_WRITE);
+    TEST_IORING_OP(IORING_OP_FADVISE);
+    TEST_IORING_OP(IORING_OP_MADVISE);
+    TEST_IORING_OP(IORING_OP_SEND);
+    TEST_IORING_OP(IORING_OP_RECV);
+    TEST_IORING_OP(IORING_OP_OPENAT2);
+    TEST_IORING_OP(IORING_OP_EPOLL_CTL);
+    TEST_IORING_OP(IORING_OP_SPLICE);
+    TEST_IORING_OP(IORING_OP_PROVIDE_BUFFERS);
+    TEST_IORING_OP(IORING_OP_REMOVE_BUFFERS);
+    TEST_IORING_OP(IORING_OP_TEE);
+    TEST_IORING_OP(IORING_OP_SHUTDOWN);
+    TEST_IORING_OP(IORING_OP_RENAMEAT);
+    TEST_IORING_OP(IORING_OP_UNLINKAT);
+    TEST_IORING_OP(IORING_OP_MKDIRAT);
+#undef TEST_IORING_OP
 
-#   define TEST_IORING_FEATURE(feature) if (p.features & feature) puts("\t" #feature)
-        puts("Supported io_uring features by current kernel:");
-        TEST_IORING_FEATURE(IORING_FEAT_SINGLE_MMAP);
-        TEST_IORING_FEATURE(IORING_FEAT_NODROP);
-        TEST_IORING_FEATURE(IORING_FEAT_SUBMIT_STABLE);
-        TEST_IORING_FEATURE(IORING_FEAT_RW_CUR_POS);
-        TEST_IORING_FEATURE(IORING_FEAT_CUR_PERSONALITY);
-        TEST_IORING_FEATURE(IORING_FEAT_FAST_POLL);
-        TEST_IORING_FEATURE(IORING_FEAT_POLL_32BITS);
-        TEST_IORING_FEATURE(IORING_FEAT_SQPOLL_NONFIXED);
-        TEST_IORING_FEATURE(IORING_FEAT_EXT_ARG);
-#   undef TEST_IORING_FEATURE
-#endif
+#define TEST_IORING_FEATURE(feature) if (p.features & feature) puts("\t" #feature)
+    puts("Supported io_uring features by current kernel:");
+    TEST_IORING_FEATURE(IORING_FEAT_SINGLE_MMAP);
+    TEST_IORING_FEATURE(IORING_FEAT_NODROP);
+    TEST_IORING_FEATURE(IORING_FEAT_SUBMIT_STABLE);
+    TEST_IORING_FEATURE(IORING_FEAT_RW_CUR_POS);
+    TEST_IORING_FEATURE(IORING_FEAT_CUR_PERSONALITY);
+    TEST_IORING_FEATURE(IORING_FEAT_FAST_POLL);
+    TEST_IORING_FEATURE(IORING_FEAT_POLL_32BITS);
+    TEST_IORING_FEATURE(IORING_FEAT_SQPOLL_NONFIXED);
+    TEST_IORING_FEATURE(IORING_FEAT_EXT_ARG);
+#undef TEST_IORING_FEATURE
     }
 
     /** Destroy io_service / io_uring object */
@@ -555,14 +557,9 @@ public:
         unsigned flags,
         uint8_t iflags = 0
     ) {
-#if LINUX_KERNEL_VERSION >= 57
         auto* sqe = io_uring_get_sqe_safe();
         io_uring_prep_splice(sqe, fd_in, off_in, fd_out, off_out, nbytes, flags);
         return await_work(sqe, iflags);
-#else
-        co_await yield(iflags);
-        co_return ::splice(fd_in, &off_in, fd_out, &off_out, nbytes, flags);
-#endif
     }
 
     /** Duplicate pipe content asynchronously
@@ -578,14 +575,14 @@ public:
         unsigned flags,
         uint8_t iflags = 0
     ) {
-#if LINUX_KERNEL_VERSION >= 58
-        auto* sqe = io_uring_get_sqe_safe();
-        io_uring_prep_tee(sqe, fd_in, fd_out, nbytes, flags);
-        return await_work(sqe, iflags);
-#else
-        co_await yield(iflags);
-        co_return ::tee(fd_in, fd_out, nbytes, flags);
-#endif
+        if (probe_ops[IORING_OP_TEE]) {
+            auto* sqe = io_uring_get_sqe_safe();
+            io_uring_prep_tee(sqe, fd_in, fd_out, nbytes, flags);
+            co_return co_await await_work(sqe, iflags);
+        } else {
+            co_await yield(iflags);
+            co_return ::tee(fd_in, fd_out, nbytes, flags);
+        }
     }
 
     /** Shut down part of a full-duplex connection asynchronously
@@ -600,14 +597,14 @@ public:
         unsigned flags,
         uint8_t iflags = 0
     ) {
-#if LINUX_KERNEL_VERSION >= 59
-        auto* sqe = io_uring_get_sqe_safe();
-        io_uring_prep_shutdown(sqe, fd, how);
-        return await_work(sqe, iflags);
-#else
-        co_await yield(iflags);
-        co_return ::shutdown(fd, how);
-#endif
+        if (probe_ops[IORING_OP_SHUTDOWN]) {
+            auto* sqe = io_uring_get_sqe_safe();
+            io_uring_prep_shutdown(sqe, fd, how);
+            co_return co_await await_work(sqe, iflags);
+        } else {
+            co_await yield(iflags);
+            co_return ::shutdown(fd, how);
+        }
     }
 
     /** Change the name or location of a file asynchronously
@@ -624,14 +621,14 @@ public:
         unsigned flags,
         uint8_t iflags = 0
     ) {
-#if LINUX_KERNEL_VERSION >= 59
-        auto* sqe = io_uring_get_sqe_safe();
-        io_uring_prep_renameat(sqe, olddfd, oldpath, newdfd, newpath, flags);
-        return await_work(sqe, iflags);
-#else
-        co_await yield(iflags);
-        co_return ::renameat2(olddfd, oldpath, newdfd, newpath, flags);
-#endif
+        if (probe_ops[IORING_OP_RENAMEAT]) {
+            auto* sqe = io_uring_get_sqe_safe();
+            io_uring_prep_renameat(sqe, olddfd, oldpath, newdfd, newpath, flags);
+            co_return co_await await_work(sqe, iflags);
+        } else {
+            co_await yield(iflags);
+            co_return ::renameat2(olddfd, oldpath, newdfd, newpath, flags);
+        }
     }
 
     /** Delete a name and possibly the file it refers to asynchronously
@@ -646,14 +643,14 @@ public:
         unsigned flags,
         uint8_t iflags = 0
     ) {
-#if LINUX_KERNEL_VERSION >= 59
-        auto* sqe = io_uring_get_sqe_safe();
-        io_uring_prep_unlinkat(sqe, dfd, path, flags);
-        return await_work(sqe, iflags);
-#else
-        co_await yield(iflags);
-        co_return ::unlinkat(dfd, path, flags);
-#endif
+        if (probe_ops[IORING_OP_UNLINKAT]) {
+            auto* sqe = io_uring_get_sqe_safe();
+            io_uring_prep_unlinkat(sqe, dfd, path, flags);
+            co_return co_await await_work(sqe, iflags);
+        } else {
+            co_await yield(iflags);
+            co_return ::unlinkat(dfd, path, flags);
+        }
     }
 
 private:
@@ -779,4 +776,5 @@ public:
 private:
     io_uring ring;
     unsigned cqe_count = 0;
+    bool probe_ops[IORING_OP_LAST] = {};
 };
